Initial dependency changes & build fixes TODO: * Update docker nginx ports * Finish moving uploader * Nginx config for production
