+++ metaspace/graphql/package.json
@@ -42,6 +42,7 @@
     "bcrypt": "^3.0.2",
     "body-parser": "^1.15.2",
     "compression": "^1.6.2",
+    "concat-files": "^0.1.1",
     "config": "^1.26.1",
     "connect-redis": "^3.3.3",
     "cors": "^2.8.1",
@@ -63,8 +64,9 @@
     "knex": "^0.19.3",
     "lodash": "^4.17.4",
     "merge-graphql-schemas": "^1.5.3",
+    "mkdirp": "^0.5.1",
     "moment": "^2.22.2",
-    "multer": "^1.3",
+    "multer": "^1.4.2",
     "node-fetch": "^1.6.3",
     "node-slack": "0.0.7",
     "passport": "^0.4.0",
