--- metaspace/webapp/.babelrc
@@ -1,22 +1,0 @@
-{
-  "env": {
-    "test": {
-      "presets": [
-        ["@babel/preset-env", {"modules": "commonjs"}]
-      ]
-    }
-  },
-  "presets": [
-    ["@babel/preset-env", {"modules": false}]
-  ],
-  "plugins": [
-    ["babel-plugin-component", {
-      "libraryName": "element-ui",
-      "styleLibraryName": "theme-default"
-    }],
-
-    "@babel/plugin-proposal-object-rest-spread",
-    "@babel/plugin-transform-runtime",
-    "@babel/plugin-syntax-dynamic-import"
-  ]
-}
