+++ metaspace/webapp/babel.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  presets: [
+    ['@vue/app', { useBuiltIns: 'entry' }]
+  ]
+}
