+++ metaspace/webapp/package.json
@@ -5,15 +5,19 @@
   "author": "Artem Tarasov <artem.tarasov@embl.de>",
   "private": true,
   "scripts": {
-    "dev": "cross-env NODE_ENV=development node --max_old_space_size=4096 server.js",
-    "build": "yarn run deref-schema && cross-env NODE_ENV=production webpack --progress --hide-modules",
-    "build-ci": "yarn run deref-schema && cross-env NODE_ENV=production webpack --hide-modules",
-    "test": "yarn run generate-local-graphql-schema && cross-env NODE_ENV=test jest",
-    "test-ci": "cross-env NODE_ENV=test jest",
-    "coverage": "cross-env NODE_ENV=test jest --coverage",
-    "deref-schema": "node ./deref_schema.js src/lib/metadataSchemas",
-    "generate-local-graphql-schema": "cd ../graphql && yarn run gen-graphql-schema",
-    "fetch-prod-graphql-schema": "apollo schema:download tests/utils/graphql-schema.json --endpoint https://metaspace2020.eu/graphql"
+    "dev": "cross-env NODE_ENV=development concurrently --raw --kill-others \"node server.js\" \"vue-cli-service serve\"",
+    "_build": "yarn run deref-schema && cross-env NODE_ENV=production webpack --progress --hide-modules",
+    "_build-ci": "yarn run deref-schema && cross-env NODE_ENV=production webpack --hide-modules",
+    "_test": "yarn run generate-local-graphql-schema && cross-env NODE_ENV=test jest",
+    "_test-ci": "cross-env NODE_ENV=test jest",
+    "_coverage": "cross-env NODE_ENV=test jest --coverage",
+    "_deref-schema": "node ./deref_schema.js src/lib/metadataSchemas",
+    "_generate-local-graphql-schema": "cd ../graphql && yarn run gen-graphql-schema",
+    "_fetch-prod-graphql-schema": "apollo schema:download tests/utils/graphql-schema.json --endpoint https://metaspace2020.eu/graphql",
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
+    "lint": "vue-cli-service lint",
+    "test:unit": "vue-cli-service test:unit"
   },
   "dependencies": {
     "@types/js-cookie": "^2.1.0",
@@ -21,16 +25,14 @@
     "@types/sanitize-html": "^1.18.2",
     "@types/upng-js": "^2.1.1",
     "@types/ws": "^5.1.2",
-    "apollo-client": "^2.6.4",
+    "@vue/cli": "^3.11.0",
     "apollo-client-preset": "^1.0.8",
     "apollo-link-batch-http": "^1.2.12",
-    "apollo-link-context": "^1.0.18",
-    "apollo-link-error": "^1.1.11",
-    "apollo-link-ws": "^1.0.18",
     "body-parser": "^1.16.0",
     "compression": "^1.6.2",
     "concat-files": "^0.1.1",
     "connect-history-api-fallback": "^1.5.0",
+    "core-js": "^3.2.1",
     "d3": "^5.11.0",
     "d3-color": "^1.0.3",
     "d3-scale": "^1.0.5",
@@ -42,11 +44,8 @@
     "express": "^4.14.0",
     "file-saver": "^1.3.3",
     "fine-uploader": "^5.13.0",
-    "fork-ts-checker-webpack-plugin": "^1.5.0",
     "graphql": "^14.5.3",
-    "graphql-tag": "^2.9.2",
     "js-cookie": "^2.2.0",
-    "json-loader": "^0.5.7",
     "lodash": "^4.17.4",
     "lodash-es": "^4.17.10",
     "marked": "^0.6.1",
@@ -60,88 +59,73 @@
     "sanitize-html": "^1.20.0",
     "serve-favicon": "^2.3.2",
     "simple-statistics": "^7.0.2",
-    "subscriptions-transport-ws": "^0.9.14",
-    "ts-loader": "^6.0.4",
+    "subscriptions-transport-ws": "^0.9.16",
     "upng-js": "https://github.com/photopea/UPNG.js.git#f550095",
     "uuid": "^3.0.1",
     "vue": "^2.6.10",
     "vue-analytics": "^5.9.1",
     "vue-apollo": "^3.0.0-rc.2",
+    "vue-class-component": "^7.1.0",
     "vue-clickaway": "^2.1.0",
     "vue-lazyload": "^1.0.0-rc5",
+    "vue-property-decorator": "^8.2.2",
     "vue-resize-directive": "^1.2.0",
-    "vue-router": "^3.0.1",
+    "vue-router": "3.0.1",
     "vue-slide-up-down": "^1.3.3",
-    "vuex": "~3.0.1",
+    "vuex": "^3.1.1",
     "vuex-router-sync": "^5.0.0",
-    "zen-observable-ts": "^0.8.9"
+    "zen-observable-ts": "^0.8.19"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/plugin-transform-runtime": "^7.0.0",
-    "@babel/preset-env": "^7.0.0",
-    "@babel/runtime": "^7.0.0",
-    "@types/cucumber": "2.0.4",
     "@types/d3": "4.10.1",
     "@types/file-saver": "0.0.1",
-    "@types/jest": "^23.1.5",
+    "@types/jest": "^24.0.18",
     "@types/lodash-es": "~4.17.0",
     "@types/node": "^8",
     "@types/node-fetch": "^2.5.0",
     "@types/plotly.js": "1.28.11",
-    "@types/webpack-env": "1.13.1",
-    "@vue/test-utils": "^1.0.0-beta.24",
-    "apollo": "^2.18.0",
+    "@vue/cli-plugin-babel": "^3.11.0",
+    "@vue/cli-plugin-typescript": "^3.11.0",
+    "@vue/cli-plugin-unit-jest": "^3.11.0",
+    "@vue/cli-service": "^3.11.0",
+    "@vue/eslint-config-typescript": "^4.0.0",
+    "@vue/test-utils": "^1.0.0-beta.29",
     "apollo-link-schema": "^1.2.3",
-    "autoprefixer": "^6.7.2",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-eslint": "^10.0.3",
     "babel-jest": "^24.9.0",
     "babel-loader": "^8.0.0",
     "babel-plugin-component": "^1.1.1",
+    "concurrently": "^4.1.2",
     "cross-env": "^3.0.0",
-    "css-loader": "^0.28.11",
-    "file-loader": "^1.1.11",
+    "eslint": "^6.3.0",
+    "eslint-plugin-vue": "^5.2.3",
     "frontmatter-markdown-loader": "^1.8.0",
-    "generate-json-webpack-plugin": "^0.3.1",
-    "graphql-tools": "^3.1.1",
-    "html-loader": "^0.5.5",
-    "jest": "^24.9.0",
-    "jest-serializer-vue": "^2.0.2",
+    "graphql-tag": "^2.10.1",
+    "graphql-tools": "^4.0.5",
     "json-schema-deref-sync": "^0.10.1",
-    "node-sass": "4.5.3",
-    "postcss-loader": "~2.1.5",
-    "sass-loader": "7.0.1",
-    "source-map-loader": "0.2.3",
-    "style-loader": "~0.21.0",
+    "node-sass": "^4.12.0",
+    "sass-loader": "^8.0.0",
     "supertest": "^3.4.2",
     "ts-jest": "^23.1.4",
-    "ts-node": "^8.3.0",
     "typescript": "^3.6.2",
-    "vue-class-component": "6.2.0",
-    "vue-jest": "^2.6.0",
-    "vue-loader": "^15.7.1",
-    "vue-property-decorator": "6.1.0",
+    "vue-cli-plugin-apollo": "^0.21.0",
     "vue-template-compiler": "^2.6.10",
-    "webpack": "^4.39.3",
-    "webpack-bundle-analyzer": "^3.3.2",
-    "webpack-cli": "^3.1.0",
-    "webpack-dev-middleware": "^3.1.3",
-    "webpack-dev-server": "^3.8.0",
-    "webpack-hot-middleware": "^2.22.2"
+    "webpack-bundle-analyzer": "^3.4.1"
   },
   "resolutions": {
     "babel-core": "7.0.0-bridge.0"
   },
   "jest": {
     "transform": {
-      "^.+\\.vue$": "<rootDir>/node_modules/vue-jest",
-      "^.+\\.[jt]sx?$": "<rootDir>/node_modules/ts-jest"
+      "^.+\\.vue$": "vue-jest",
+      ".+\\.(css|styl|less|sass|scss|svg|png|jpg|ttf|woff|woff2)$": "jest-transform-stub",
+      "^.+\\.tsx?$": "ts-jest"
     },
     "moduleNameMapper": {
       "lodash-es": "lodash",
       "\\.scss$": "<rootDir>/tests/utils/scssMock.js",
-      "\\.(png|jpg|gif|svg)$": "<rootDir>/tests/utils/staticFileMock.js"
+      "^@/(.*)$": "<rootDir>/src/$1"
     },
     "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$",
     "setupFilesAfterEnv": [
@@ -149,7 +133,7 @@
     ],
     "globals": {
       "ts-jest": {
-        "tsConfig": "tsconfig.spec.json"
+        "babelConfig": true
       }
     },
     "clearMocks": true,
@@ -165,12 +149,37 @@
       "!src/**/*.{test,spec,d}.{ts,js}"
     ],
     "snapshotSerializers": [
-      "<rootDir>/node_modules/jest-serializer-vue"
+      "jest-serializer-vue"
+    ],
+    "testURL": "http://localhost/",
+    "watchPlugins": [
+      "jest-watch-typeahead/filename",
+      "jest-watch-typeahead/testname"
     ]
   },
   "browserslist": {
     "production": [
       "defaults"
     ]
+  },
+  "_eslintConfig": {
+    "root": true,
+    "env": {
+      "node": true
+    },
+    "extends": [
+      "plugin:vue/essential",
+      "eslint:recommended",
+      "@vue/typescript"
+    ],
+    "rules": {},
+    "parserOptions": {
+      "parser": "@typescript-eslint/parser"
+    }
+  },
+  "postcss": {
+    "plugins": {
+      "autoprefixer": {}
+    }
   }
 }
